+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixes
 - Fix rotation locks
 - `--force-without-root` message fix at startup
+- Fix standard responses if initial check on log request failed
 
 ## [1.5.0] - 23-02-2018
 ### Updates